ide: Fix a rare hang during block draining
If the guest issues a discard during a block drain section, the
blk_aio_pdiscard() may not be processed, but queued instead.
And so the callback will never be called to issue the bh and decrease
the BB in-flight number again.
This causes a hang in the drain code, since it will wait forever for the
BB in-flight counter to decrease.

This reverts commit 7e5cdb34 "ide: Increment BB in-flight counter for TRIM BH"
to fix this hang.

The bug that the reverted commit fixed will be handled in the next commit.

Signed-off-by: Lukas Straub <lukasstraub2@web.de>